Change helper to support ruclei
The idea is to create issues which have check: "ruclei/template_file.yaml". The helper that starts tasks will then check for "ruclei" in the name of the task, and run the hardcoded task named "ruclei_runner" if its found. To tell the ruclei which template to run, the helper extracts the temlate file name from the check, and passes it as a user option.